backend/chat_orchestrator/requirements.txt
fastapi
uvicorn[standard]
httpx
jinja2
pydantic
python-dotenv
